вроде работает хосподь помоги

+++ vk_bot/mods/other/speechtotext.py
@@ -3,6 +3,8 @@
 """
 from vk_bot.core.modules.basicplug import BasicPlug
 from vk_bot.core.modules.upload import Upload
+from vk_bot.core.sql.vksql import *
+from loadevn import speechtotext
 import speech_recognition as sr
 from pydub import AudioSegment
 import os
@@ -16,25 +18,28 @@ class AutdiotoText(BasicPlug, Upload):
     attachment = 'audio_message'
 
     def main(self):
-        try:
-            link = self.event.object["attachments"][0]["audio_message"]["link_mp3"]
-        except:
+        if checktable(f'{speechtotext}', "chat_id", f"{self.event.object.peer_id}")['status'] == 0:
             return
-        name = self.dowloadfile(link)['name']
-        sound = AudioSegment.from_mp3(name)
-        soundname = f"audio2{time.time_ns()}.wav"
-        sound.export(soundname, format="wav")
+        else:
+            try:
+                link = self.event.object["attachments"][0]["audio_message"]["link_mp3"]
+            except:
+                return
+            name = self.dowloadfile(link)['name']
+            sound = AudioSegment.from_mp3(name)
+            soundname = f"audio2{time.time_ns()}.wav"
+            sound.export(soundname, format="wav")
 
-        r = sr.Recognizer()  # Использование файла как источник
-        with sr.AudioFile(soundname) as source:
-            audio = r.record(source)  # Считывает весь файл
-        try:
-            result = r.recognize_google(audio, language="ru_RU")
-            self.sendmsg(f"сказал: {result}")
-        except sr.UnknownValueError:
-            return
-        except sr.RequestError:
-            self.sendmsg(f"Ошибка при отправки запроса")
-        finally:
-            os.remove(name)
-            os.remove(soundname)
+            r = sr.Recognizer()  # Использование файла как источник
+            with sr.AudioFile(soundname) as source:
+                audio = r.record(source)  # Считывает весь файл
+            try:
+                result = r.recognize_google(audio, language="ru_RU")
+                self.sendmsg(f"сказал: {result}")
+            except sr.UnknownValueError:
+                return
+            except sr.RequestError:
+                self.sendmsg(f"Ошибка при отправки запроса")
+            finally:
+                os.remove(name)
+                os.remove(soundname)
